refractor login service
