+++ src/main/java/com/ffc/ffc_be/controller/LoginController.java
@@ -3,8 +3,7 @@ package com.ffc.ffc_be.controller;
 import com.ffc.ffc_be.model.builder.ResponseDto;
 import com.ffc.ffc_be.model.dto.request.LoginRequest;
 import com.ffc.ffc_be.model.dto.response.LoginResponse;
-import com.ffc.ffc_be.service.AuthService;
-import jakarta.validation.Valid;
+import com.ffc.ffc_be.security.AuthenticationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,10 +15,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/login")
 @RequiredArgsConstructor
 public class LoginController {
-    private final AuthService authService;
+    private final AuthenticationService authenticationService;
 
     @PostMapping
     public ResponseEntity<ResponseDto<LoginResponse>> authenticate(@RequestBody LoginRequest request) {
-        return authService.requestLogin(request);
+        return authenticationService.requestLogin(request);
     }
 }
