+++ src/main/java/com/ffc/ffc_be/security/AuthenticationService.java
@@ -1,10 +1,15 @@
 package com.ffc.ffc_be.security;
 
+import com.ffc.ffc_be.model.builder.ResponseBuilder;
+import com.ffc.ffc_be.model.builder.ResponseDto;
 import com.ffc.ffc_be.model.dto.request.LoginRequest;
 import com.ffc.ffc_be.model.dto.request.RegisterRequest;
+import com.ffc.ffc_be.model.dto.response.LoginResponse;
 import com.ffc.ffc_be.model.entity.UserInfoModel;
+import com.ffc.ffc_be.model.enums.StatusCodeEnum;
 import com.ffc.ffc_be.repository.IUserInfoRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -18,6 +23,8 @@ public class AuthenticationService {
     private final PasswordEncoder passwordEncoder;
 
     private final AuthenticationManager authenticationManager;
+
+    private final JwtService jwtService;
 
     public UserDetailsImpl signup(RegisterRequest request) {
         UserInfoModel user = UserInfoModel.builder()
@@ -40,5 +47,23 @@ public class AuthenticationService {
 
         return new UserDetailsImpl(userRepository.findByUsername(request.getUsername())
                 .orElseThrow());
+    }
+
+    public ResponseEntity<ResponseDto<LoginResponse>> requestLogin(LoginRequest request) {
+        UserDetailsImpl authenticatedUser;
+        try {
+            authenticatedUser = authenticate(request);
+        } catch (Exception e) {
+            return ResponseBuilder.badRequestResponse("Login Failed", StatusCodeEnum.STATUSCODE2001);
+        }
+
+        String jwtToken = jwtService.generateToken(authenticatedUser);
+
+        LoginResponse loginResponse = LoginResponse.builder()
+                .token(jwtToken)
+                .expiresIn(jwtService.getExpirationTime())
+                .build();
+
+        return ResponseBuilder.okResponse("Login successfully", loginResponse, StatusCodeEnum.STATUSCODE1001);
     }
 }
