--- src/main/java/com/ffc/ffc_be/security/UnauthorizedEntryPoint.java
@@ -1,19 +1,0 @@
-package com.ffc.ffc_be.security;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.io.Serializable;
-
-@Component
-public class UnauthorizedEntryPoint implements AuthenticationEntryPoint, Serializable {
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthenticated");
-    }
-}
