+++ src/main/java/com/ffc/ffc_be/service/UserInfoService.java
@@ -1,14 +1,29 @@
 package com.ffc.ffc_be.service;
 
+import com.ffc.ffc_be.model.builder.ResponseBuilder;
+import com.ffc.ffc_be.model.builder.ResponseDto;
+import com.ffc.ffc_be.model.dto.request.RegisterRequest;
 import com.ffc.ffc_be.model.entity.UserInfoModel;
+import com.ffc.ffc_be.model.enums.StatusCodeEnum;
+import com.ffc.ffc_be.security.AuthenticationService;
 import com.ffc.ffc_be.security.UserDetailsImpl;
+import com.ffc.ffc_be.security.UserDetailsServiceImpl;
+import jakarta.validation.Valid;
 import lombok.CustomLog;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestBody;
 
 @Service
 @CustomLog
+@RequiredArgsConstructor
 public class UserInfoService {
+    private final AuthenticationService authService;
+    private final UserDetailsServiceImpl userDetailsService;
+
     public UserInfoModel getUserInfoFromContext() {
         try {
             UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
@@ -19,5 +34,25 @@ public class UserInfoService {
 
             return null;
         }
+    }
+
+    public ResponseEntity<ResponseDto<UserInfoModel>> register(@RequestBody @Valid RegisterRequest request) {
+        UserDetailsImpl registeredUser;
+        try {
+            UserDetails usernameExist = userDetailsService.loadUserByUsername(request.getUsername());
+            if (usernameExist != null) {
+                return ResponseBuilder.badRequestResponse("Create new user failed, please use other username",
+                        StatusCodeEnum.STATUSCODE2001);
+            }
+
+            registeredUser = authService.signup(request);
+        } catch (Exception e) {
+            return ResponseBuilder.badRequestResponse("Create new user failed, unexpected error",
+                    StatusCodeEnum.STATUSCODE2001);
+        }
+
+        return ResponseBuilder.okResponse("Create new user successfully",
+                registeredUser.getUser(),
+                StatusCodeEnum.STATUSCODE1001);
     }
 }
